backend/requirements.txt

fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pytest
